Added Refresh Option MenuItem
This builds out the left panel with a button currently (might change to
textfield). Formatting not currently in place. However tracking what is
saved and not saved is currently working.
Can create a refresh function that will be used to build the panel and
be called at multiple points of the App.